#1994 Secure DWR methods - correct security dwr init/save/delete

+++ src/com/serotonin/mango/web/dwr/PointLinksDwr.java
@@ -18,13 +18,6 @@
  */
 package com.serotonin.mango.web.dwr;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import javax.script.ScriptException;
-
 import com.serotonin.db.IntValuePair;
 import com.serotonin.mango.Common;
 import com.serotonin.mango.db.dao.DataPointDao;
@@ -35,7 +28,6 @@ import com.serotonin.mango.rt.dataImage.PointValueTime;
 import com.serotonin.mango.rt.dataSource.meta.ResultTypeException;
 import com.serotonin.mango.rt.dataSource.meta.ScriptExecutor;
 import com.serotonin.mango.rt.link.PointLinkRT;
-import com.serotonin.mango.util.LoggingScriptUtils;
 import com.serotonin.mango.vo.DataPointExtendedNameComparator;
 import com.serotonin.mango.vo.DataPointVO;
 import com.serotonin.mango.vo.User;
@@ -48,6 +40,12 @@ import com.serotonin.web.taglib.DateFunctions;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
+import javax.script.ScriptException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import static com.serotonin.mango.util.LoggingScriptUtils.infoErrorExecutionScript;
 
 /**
@@ -57,6 +55,7 @@ public class PointLinksDwr extends BaseDwr {
     private static final Log LOG = LogFactory.getLog(PointLinksDwr.class);
     public Map<String, Object> init() {
         User user = Common.getUser();
+        Permissions.ensureAdmin(user);
         Map<String, Object> data = new HashMap<String, Object>();
 
         // Get the points that this user can access.
@@ -109,6 +108,7 @@ public class PointLinksDwr extends BaseDwr {
     public DwrResponseI18n savePointLink(int id, String xid, int sourcePointId, int targetPointId, String script,
             int event, boolean disabled) {
         // Validate the given information. If there is a problem, return an appropriate error message.
+        Permissions.ensureAdmin();
         PointLinkVO vo = new PointLinkVO();
         vo.setId(id);
         vo.setXid(xid);
@@ -138,6 +138,7 @@ public class PointLinksDwr extends BaseDwr {
     }
 
     public void deletePointLink(int id) {
+        Permissions.ensureAdmin();
         Common.ctx.getRuntimeManager().deletePointLink(id);
     }
